Hide C-string tests from old toolchain versions
