+++ .github/workflows/ci.yml
@@ -24,7 +24,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        rust: [nightly, beta, stable, 1.76.0, 1.60.0]
+        rust: [nightly, beta, stable, 1.77.0, 1.76.0, 1.60.0]
     timeout-minutes: 45
     steps:
       - uses: actions/checkout@v5
